+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Constellation Changes
 
+## Changes in January 2022
+
+-   Added a new parameter `tabCaption` in `newTab` public methods in `DataAccessTabPane` class 
+	to provide the Step tab caption when required. This is used when the user renames the 
+	default caption.	
+	
 ## Changes in December 2021
 
 -   Changed the return type of `processVertices` and `processTransactions` methods